Refactor import_params to be its own utility
